feat: authentication implemented

+++ database/migration/000003_create_feedbacks_table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE feedbacks
 (
-    id BIGINT PRIMARY KEY NOT NULL,
+    id SERIAL PRIMARY KEY NOT NULL,
     email VARCHAR(255) NOT NULL,
     feedback_text VARCHAR(255) NOT NULL,
     source VARCHAR(255) NOT NULL
